Organize tsx dir
